add explicit integer

See merge request Oslandia/qgis/ign-geotuileur!114

+++ geotuileur/gui/tile_creation/qwp_tile_generation_status.py
@@ -210,7 +210,7 @@ class TileGenerationStatusPageWizard(QWizardPage):
 
         # Run timer for tile creation check
         self.create_tile_check_timer.start(
-            PlgOptionsManager.get_plg_settings().status_check_sleep / 1000.0
+            int(PlgOptionsManager.get_plg_settings().status_check_sleep / 1000.0)
         )
 
     def _create_tile_finished(self, context, successful, results):
